\name{Acces}
\alias{Acces}
\title{Access to an element of a list}
\description{To access to an element of a matrix in a list of matrices 
  
}
\usage{Acces(mat,i,j)

}
\arguments{
  \item{mat      }{Matrix.  }
  \item{i       }{Row of the matrix. }
  \item{j   }{Column of the matrix.  }
}
\details{
  
}
\value{
  The value of the element mat[i,j].
}
\references{ 

 }

\seealso{}

\keyword{internal }
