\name{CalculDirect}
\alias{CalculDirect}
\title{ Estimator of D }
\description{
  Compute the estimator of the orientation matrix D which is the same for the two first components.  
}
\usage{
  CalculDirect(W)
}
\arguments{
  \item{W  }{List of the K dispersion matrices.  }
}
\details{
  
}
\value{
  Estimator of the D matrix which has same orientation for the two first components.
}
\references{ 

 }

\seealso{matW}

\keyword{internal}

