\name{Initialization}
\alias{Initialization}
\title{ Random initialization of the posterior probabilities }
\description{
  Random initialization of the posterior probabilities
}
\usage{
  Initialization(n,K)
}
\arguments{
  \item{n   }{Number of observations.  }
  \item{K   }{Number of groups.  }

}

\value{
  A matrix (n*K) of the posterior probabilities with 0 or 1.
}

\keyword{internal }
