\name{ProdMat}
\alias{ProdMat}
\title{Compute a product of matrices  }
\description{Compute the product between A transpose, B and A.
  
}
\usage{ProdMat(A,B)

}
\arguments{
  \item{A       }{Matrix  }
  \item{B       }{Matrix  }
  
}
\details{
  
}
\value{
  Return the value of the product between A transpose, B and A.
}
\references{ 

 }

\seealso{}

\keyword{internal}

