\name{ProdMat2}
\alias{ProdMat2}
\title{Compute a product of matrices  }
\description{Compute the product between A,B and A transpose.
  
}
\usage{ProdMat2(A,B)

}
\arguments{
  \item{A       }{Matrix  }
  \item{B       }{Matrix  }
}

\value{
  Return the value of the product between A,B and A transpose.
}


\keyword{internal}

