\name{ProdMatGauss}
\alias{ProdMatGauss}
\title{Compute a product of matrices  }
\description{
  Compute the product between -(1/2),A,B and A transpose.
}
\usage{ProdMatGauss(A,B)

}
\arguments{
  \item{A       }{Matrix.  }
  \item{B       }{Matrix.   }
}

\value{
  Return the value of the product with A, B, A transpose and -0.5.
}


\keyword{internal}
