\name{Sort_file}
\alias{Sort_file}
\title{Sort the input file  }
\description{
  Sort the file according to the column named ID which contains the positions of the probes along the chromosome.
}
\usage{
   Sort_file(fileIN,fileOUTdata)
}
\arguments{
  \item{fileIN       }{The name of the input file that we want to sort  }
  \item{fileOUTdata  }{The name of the output file sorted  }
 
}
\details{
  The input file must contain a column named ID that the file is sorted.
}
\value{
  The output file is sorted according to the position along the chromosome.
}
\references{ 

 }

\seealso{}

\keyword{internal}
